update20181216
20181216 V1.0 正式版---
修复SQLServer延时盲注paylaod缺少and导致语句错误无法获取数据的问题。
修复优化自动识别注入漏报和选择类型错误的问题。
修复bool延时注入时，获取数据判断数据大小方式不正确的问题。
修复错误显示注入获取数据时，数据可能被HTML编码或转义到导致数据不正确的问题。
修复延时注入在正常页面响应速度非常快的情况下，由于计数器可能有误差，导致判断值不正确，而无法正确获取数据（设置了20*time毫秒的误差值）。
修复读取文件，执行命令无法使用延时判断问题。
优化配置文件。
新增支持注入PostgreSQL，目前可获取数据，执行命令和文件操作稍后版本更新。

+++ SuperSQLInjection/model/ServerInfo.cs
@@ -23,7 +23,7 @@ namespace model
         public int length = 0;
         public int code = 0;
         public int location = 0;
-        public int runTime = 0;//获取网页消耗时间，毫秒
+        public long runTime = 0;//获取网页消耗时间，毫秒
         public int sleepTime = 0;//休息时间
         public String cookies = "";
         public Boolean timeout = false;
